requirements.txt
feedparser==6.0.11
openai>=1.0.0
python-dotenv==1.1.0
requests==2.32.3
slack-sdk==3.35.0
